swaybar: move i3bar definitions into separate file
