+++ swaybar/i3bar.c
@@ -6,7 +6,9 @@
 #include <string.h>
 #include <unistd.h>
 #include <wlr/util/log.h>
+#include "swaybar/bar.h"
 #include "swaybar/config.h"
+#include "swaybar/i3bar.h"
 #include "swaybar/status_line.h"
 
 void i3bar_block_unref(struct i3bar_block *block) {
